Seperation for get accounts functions
